1) Добавил обновление машин 2) Обновил entity Spares, Works, Spares 3) Убрал лишнее из контракта app.yml 4) Добавил глобальные обработчики ошибок 5) В CarController доделал все контроллеры

+++ app/src/main/java/ru/aservice/app/service/api/CarService.java
@@ -6,9 +6,7 @@ import java.util.List;
 
 public interface CarService {
 
-    //TODO: привести нормальные типы возвращаемых объектов
-
-    void saveCar(CarDto dto, Integer carId);
+    void saveCar(CarDto dto, Long carId);
 
     void deleteCar(Long carId);
